+++ client/src/components/dashboard/ProgramForm/ProgramForm.jsx
@@ -1,80 +1,32 @@
 import { useEffect, useMemo, useRef, useState } from 'react';
 
 import {
-  Box,
   Button,
   Center,
-  Checkbox,
   Drawer,
   DrawerBody,
   DrawerCloseButton,
   DrawerContent,
   DrawerHeader,
   DrawerOverlay,
-  FormControl,
-  FormLabel,
-  Heading,
   HStack,
-  Input,
-  Select,
   Spinner,
   useDisclosure,
   VStack,
 } from '@chakra-ui/react';
 
 import { MediaUploadModal } from '@/components/media/MediaUploadModal';
-import { PartnerOrganizationField } from '@/components/partners/PartnerOrganizationField';
 import { useAuthContext } from '@/contexts/hooks/useAuthContext';
 import { useBackendContext } from '@/contexts/hooks/useBackendContext';
 import ISO6391 from 'iso-639-1';
 import { useTranslation } from 'react-i18next';
 
-import { AssignedDirectorsSection } from './AssignedDirectorsSection';
-import { LocationLanguageSection } from './LocationLanguageSection';
-import { MediaPreviewTag } from './MediaPreviewTag';
-import { ResourcesSection } from './ResourcesSection';
-import { StudentsInstrumentsSection } from './StudentsInstrumentsSection';
+import { isPdfByType } from './programFormHelpers';
+import { ProgramFormMediaTab } from './ProgramFormMediaTab';
+import { ProgramFormOverviewTab } from './ProgramFormOverviewTab';
+import { saveProgramForm } from './programFormSave';
+import { useProgramFormLoad } from './useProgramFormLoad';
 
-/** Deletes instrument_change rows for this program + instrument (same as ProgramUpdateForm). */
-async function deleteInstrumentChangesForProgramInstrument(
-  backend,
-  programId,
-  instrumentId
-) {
-  const numericProgramId = Number(programId);
-  const numericInstrumentId = Number(instrumentId);
-  if (Number.isNaN(numericInstrumentId)) return;
-
-  const res = await backend.get('/program-updates');
-  const updates = Array.isArray(res.data) ? res.data : [];
-  const updateIds = updates
-    .filter((u) => Number(u.programId ?? u.program_id) === numericProgramId)
-    .map((u) => u.id);
-
-  for (const updateId of updateIds) {
-    const icRes = await backend.get(`/instrument-changes/update/${updateId}`);
-    const rows = Array.isArray(icRes.data) ? icRes.data : [];
-    for (const row of rows) {
-      const iid = Number(row.instrumentId ?? row.instrument_id);
-      if (
-        iid === numericInstrumentId &&
-        row.id !== null &&
-        row.id !== undefined
-      ) {
-        await backend.delete(`/instrument-changes/${row.id}`);
-      }
-    }
-  }
-}
-function removeFormItemByIdOrKey(list, item) {
-  return (list ?? []).filter((m) => {
-    const itemHasId = item.id !== null && item.id !== undefined;
-    const mHasId = m.id !== null && m.id !== undefined;
-    if (itemHasId && mHasId) return Number(m.id) !== Number(item.id);
-    if (item.s3_key && m.s3_key) return m.s3_key !== item.s3_key;
-    return m !== item;
-  });
-}
 export const ProgramForm = ({
   isOpen: isOpenProp,
   onOpen: onOpenProp,
@@ -139,284 +91,17 @@ export const ProgramForm = ({
   const [activeTab, setActiveTab] = useState('overview');
   const [isLoadingProgramData, setIsLoadingProgramData] = useState(false);
 
-  useEffect(() => {
-    let cancelled = false;
-
-    async function loadProgramRegionData() {
-      if (!program) {
-        setIsLoadingProgramData(false);
-        setFormState({
-          status: null,
-          programName: null,
-          partnerOrg: null,
-          showPartnerOrgOnMap: false,
-          launchDate: null,
-          regionId: null,
-          country: null,
-          countryIsoCode: null,
-          state: null,
-          city: null,
-          students: 0,
-          graduatedStudents: 0,
-          instruments: {},
-          languages: [],
-          programDirectors: [],
-          curriculumLinks: [],
-          media: [],
-          fileChanges: [],
-        });
-        setInitialProgramDirectorIds([]);
-        setInitialInstrumentQuantities({});
-        setInitialCurriculumLinks([]);
-        setInitialUploadedMedia([]);
-        setInitialGraduated(0);
-        return;
-      }
-
-      setIsLoadingProgramData(true);
-
-      let record = program;
-      try {
-        const { data } = await backend.get(`/program/${program.id}`);
-        record = {
-          ...program,
-          ...data,
-          programDirectors: program.programDirectors ?? [],
-          regionalDirectors: program.regionalDirectors ?? [],
-          playlists: program.playlists ?? [],
-          media: program.media ?? [],
-          fileChanges: program.fileChanges ?? [],
-        };
-      } catch (err) {
-        console.error('ProgramForm: could not load program details', err);
-      }
-
-      try {
-        const mediaRes = await backend.get(`/program/${program.id}/media`);
-        if (Array.isArray(mediaRes.data)) {
-          record = { ...record, media: mediaRes.data };
-        }
-      } catch (err) {
-        console.error('ProgramForm: could not load program media', err);
-      }
-
-      try {
-        const filesRes = await backend.get(
-          `/fileChanges/program/${program.id}`
-        );
-        if (Array.isArray(filesRes.data)) {
-          record = { ...record, fileChanges: filesRes.data };
-        }
-      } catch (err) {
-        console.error('ProgramForm: could not load program file changes', err);
-      }
-
-      try {
-        const pdRes = await backend.get(
-          `/program/${program.id}/program-directors`
-        );
-        if (Array.isArray(pdRes.data)) {
-          record = { ...record, programDirectors: pdRes.data };
-        }
-      } catch (err) {
-        console.error('ProgramForm: could not load program directors', err);
-      }
-
-      let regionId = null;
-      let countryIsoCode = null;
-
-      if (record.country) {
-        try {
-          const countryResponse = await backend.get(
-            `/country/${record.country}`
-          );
-          regionId = countryResponse.data.regionId;
-          const rawIso =
-            countryResponse.data.isoCode ?? countryResponse.data.iso_code;
-          if (
-            rawIso !== null &&
-            rawIso !== undefined &&
-            String(rawIso).trim() !== ''
-          ) {
-            countryIsoCode = String(rawIso).trim().toUpperCase();
-          }
-        } catch (error) {
-          console.error('error fetching country/region', error);
-        }
-      }
-
-      const mappedProgramDirectors = (record.programDirectors ?? []).map(
-        (d) => ({
-          userId: d.userId ?? d.id ?? d.user_id,
-          firstName: d.firstName,
-          lastName: d.lastName,
-          picture: d.picture,
-        })
-      );
-      const programDirectorsForForm = mappedProgramDirectors.slice(0, 1);
-
-      const instrumentMap = {};
-      const initialInstrumentMap = {};
-      try {
-        const instrumentsResponse = await backend.get(
-          `/program/${record.id}/instruments`
-        );
-        const instruments = instrumentsResponse.data || [];
-        instruments.forEach((inst) => {
-          const id = inst.instrumentId ?? inst.id;
-          if (!id) return;
-          const qty = inst.quantity ?? 0;
-          if (qty === 0) return;
-          instrumentMap[id] = {
-            id,
-            name: inst.name,
-            quantity: qty,
-          };
-          initialInstrumentMap[id] = qty;
-        });
-      } catch (err) {
-        console.error('Error fetching program instruments:', err);
-      }
-
-      const normalizedLanguages = (() => {
-        if (Array.isArray(record.languages)) {
-          return [
-            ...new Set(
-              record.languages
-                .map((value) => String(value).trim().toLowerCase())
-                .filter((value) => ISO6391.validate(value))
-            ),
-          ];
-        }
-        const existingValue = record.primaryLanguage;
-        if (!existingValue) return [];
-        const trimmedValue = String(existingValue).trim();
-        if (ISO6391.validate(trimmedValue.toLowerCase())) {
-          return [trimmedValue.toLowerCase()];
-        }
-        const mappedCode = ISO6391.getCode(trimmedValue);
-        return mappedCode ? [mappedCode.toLowerCase()] : [];
-      })();
-
-      let graduatedTotal = 0;
-      let sumEnrollment = 0;
-      try {
-        const aggRes = await backend.get(
-          `/program/${record.id}/enrollment-aggregates`
-        );
-        graduatedTotal = Number(aggRes.data?.sumGraduated ?? 0);
-        sumEnrollment = Number(aggRes.data?.sumEnrollment ?? 0);
-      } catch (error) {
-        console.error('Error fetching enrollment aggregates:', error);
-      }
-      if (Number.isNaN(graduatedTotal)) graduatedTotal = 0;
-      if (Number.isNaN(sumEnrollment)) sumEnrollment = 0;
-      setInitialGraduated(graduatedTotal);
-
-      const netStudentsFromAgg = sumEnrollment - graduatedTotal;
-      const resolvedStudents =
-        record.students !== null && record.students !== undefined
-          ? Number(record.students)
-          : netStudentsFromAgg;
-
-      setFormState({
-        status: record.status ?? null,
-        programName: record.title ?? '',
-        partnerOrg: record.partnerOrg ?? record.partner_org ?? null,
-        showPartnerOrgOnMap:
-          record.showPartnerOrgOnMap ?? record.show_partner_org_on_map ?? false,
-        launchDate: record.launchDate ? record.launchDate.split('T')[0] : '',
-        regionId: regionId,
-        state: record.state ?? null,
-        city:
-          record.city !== null &&
-          record.city !== undefined &&
-          record.city !== ''
-            ? Number(record.city)
-            : null,
-        country: record.country ?? null,
-        countryIsoCode,
-        students: Number.isNaN(resolvedStudents) ? 0 : resolvedStudents,
-        graduatedStudents: graduatedTotal,
-        instruments: instrumentMap,
-        languages: normalizedLanguages,
-
-        programDirectors: programDirectorsForForm,
-
-        curriculumLinks: Array.isArray(record.playlists)
-          ? record.playlists
-              .filter(
-                (p) =>
-                  p.link &&
-                  ((p.instrumentId !== null && p.instrumentId !== undefined) ||
-                    (p.instrument_id !== null && p.instrument_id !== undefined))
-              )
-              .map((p) => ({
-                link: p.link,
-                name: p.name || 'Playlist',
-                instrumentId: p.instrumentId ?? p.instrument_id,
-                instrumentName: p.instrumentName ?? p.instrument_name ?? '',
-              }))
-          : [],
-
-        media: Array.isArray(record.media)
-          ? record.media.map((m) => ({
-              id: m.id,
-              s3_key: m.s3_key,
-              file_name: m.file_name,
-              file_type: m.file_type,
-            }))
-          : [],
-
-        fileChanges: Array.isArray(record.fileChanges)
-          ? record.fileChanges.map((f) => ({
-              id: f.id,
-              s3_key: f.s3_key,
-              file_name: f.file_name,
-              file_type: f.file_type,
-            }))
-          : [],
-      });
-
-      setInitialProgramDirectorIds(
-        mappedProgramDirectors.map((d) => d.userId).filter(Boolean)
-      );
-      setInitialInstrumentQuantities(initialInstrumentMap);
-      setInitialCurriculumLinks(
-        (record.playlists ?? [])
-          .filter(
-            (p) =>
-              p.link &&
-              ((p.instrumentId !== null && p.instrumentId !== undefined) ||
-                (p.instrument_id !== null && p.instrument_id !== undefined))
-          )
-          .map((p) => ({
-            link: p.link,
-            instrumentId: p.instrumentId ?? p.instrument_id,
-          }))
-      );
-
-      setInitialUploadedMedia(
-        [...(record.media ?? []), ...(record.fileChanges ?? [])]
-          .filter((m) => m.file_name)
-          .map((m) => m.file_name)
-      );
-    }
-
-    (async () => {
-      try {
-        await loadProgramRegionData();
-      } finally {
-        if (!cancelled) {
-          setIsLoadingProgramData(false);
-        }
-      }
-    })();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [program, backend]);
+  useProgramFormLoad({
+    program,
+    backend,
+    setFormState,
+    setInitialProgramDirectorIds,
+    setInitialInstrumentQuantities,
+    setInitialCurriculumLinks,
+    setInitialUploadedMedia,
+    setInitialGraduated,
+    setIsLoadingProgramData,
+  });
 
   useEffect(() => {
     if (isOpen) {
@@ -440,10 +125,6 @@ export const ProgramForm = ({
     setFormState((prev) => ({ ...prev, languages: languageChanges }));
   }
 
-  function isPdfByType(media) {
-    return media.file_type === 'application/pdf';
-  }
-
   const handleUploadComplete = (newFiles) => {
     if (mediaUploadTargetRef.current === 'files') {
       const pdfs = newFiles.filter((f) => isPdfByType(f));
@@ -465,253 +146,19 @@ export const ProgramForm = ({
 
   async function handleSave() {
     try {
-      const rawPartner = formState.partnerOrg;
-      const hasPartnerOrg =
-        rawPartner !== null && rawPartner !== undefined && rawPartner !== '';
-      const partnerOrgPayload = hasPartnerOrg
-        ? Number(rawPartner)
-        : program
-          ? null
-          : 1;
-
-      const data = {
-        name: formState.programName,
-        title: formState.programName,
-        status: formState.status,
-        launchDate: formState.launchDate,
-        country: formState.country,
-        state: formState.state,
-        city: formState.city,
-        students: formState.students ?? 0,
-        languages: formState.languages ?? [],
-        partnerOrg: partnerOrgPayload,
-        createdBy: currentUser?.uid || currentUser?.id,
-        description: '',
-      };
-
-      let programId;
-      const oldStudentCount = program?.students || 0;
-      const deltaNet = (formState.students ?? 0) - oldStudentCount;
-      const deltaGrad = (formState.graduatedStudents ?? 0) - initialGraduated;
-      const enrollmentDelta = deltaNet + deltaGrad;
-      const graduatedDelta = deltaGrad;
-
-      if (program) {
-        await backend.put(`/program/${program.id}`, data);
-        programId = program.id;
-      } else {
-        const response = await backend.post(`/program`, data);
-        programId = response.data.id;
-      }
-
-      const selectedDirector = formState.programDirectors[0];
-      const selectedRaw = selectedDirector?.userId;
-      const selectedKey =
-        selectedRaw !== null &&
-        selectedRaw !== undefined &&
-        String(selectedRaw).trim() !== ''
-          ? String(selectedRaw).trim()
-          : '';
-      const validSelected = selectedKey !== '';
-
-      const initialIds = initialProgramDirectorIds || [];
-      const initialKeySet = new Set(
-        initialIds.map((id) =>
-          id !== null && id !== undefined ? String(id).trim() : ''
-        )
-      );
-
-      if (validSelected) {
-        for (const uid of initialIds) {
-          if (String(uid).trim() === selectedKey) continue;
-          await backend.delete(
-            `/program-directors/${encodeURIComponent(String(uid))}`,
-            {
-              params: { programId },
-            }
-          );
-        }
-        if (!initialKeySet.has(selectedKey)) {
-          await backend.post(`/program-directors`, {
-            userId: selectedRaw,
-            programId,
-          });
-        }
-        setInitialProgramDirectorIds([selectedRaw]);
-      } else {
-        for (const uid of initialIds) {
-          await backend.delete(
-            `/program-directors/${encodeURIComponent(String(uid))}`,
-            {
-              params: { programId },
-            }
-          );
-        }
-        setInitialProgramDirectorIds([]);
-      }
-
-      const currentLinkKeys = (formState.curriculumLinks ?? []).map(
-        (p) => `${p.link}\0${p.instrumentId}`
-      );
-      for (const playlist of formState.curriculumLinks ?? []) {
-        const key = `${playlist.link}\0${playlist.instrumentId}`;
-        if (
-          !initialCurriculumLinks.some(
-            (i) => `${i.link}\0${i.instrumentId}` === key
-          )
-        ) {
-          await backend.post(`/program/${programId}/playlists`, {
-            link: playlist.link,
-            name: playlist.name || 'Playlist',
-            instrumentId: playlist.instrumentId,
-          });
-        }
-      }
-      for (const old of initialCurriculumLinks) {
-        if (!currentLinkKeys.includes(`${old.link}\0${old.instrumentId}`)) {
-          await backend.delete(`/program/${programId}/playlists`, {
-            data: { link: old.link, instrumentId: old.instrumentId },
-          });
-        }
-      }
-
-      const currentMediaIds = formState.media
-        .map((m) => m.id)
-        .filter((id) => id !== undefined && id !== null);
-      const currentFileIds = formState.fileChanges
-        .map((f) => f.id)
-        .filter((id) => id !== undefined && id !== null);
-
-      const pendingMedia = formState.media.filter((m) => !m.id);
-      const pendingFiles = formState.fileChanges.filter((f) => !f.id);
-
-      if (program) {
-        const programMedia = program?.media ?? [];
-        const mediaToDelete = programMedia.filter(
-          (oldMedia) =>
-            oldMedia.id !== null &&
-            oldMedia.id !== undefined &&
-            !currentMediaIds.includes(oldMedia.id)
-        );
-
-        for (const media of mediaToDelete) {
-          if (media.id) {
-            await backend.delete(`/mediaChange/${media.id}`);
-          }
-        }
-
-        const programFiles = program?.fileChanges ?? [];
-        const filesToDelete = programFiles.filter(
-          (oldFile) =>
-            oldFile.id !== null &&
-            oldFile.id !== undefined &&
-            !currentFileIds.includes(oldFile.id)
-        );
-
-        for (const file of filesToDelete) {
-          if (file.id) {
-            await backend.delete(`/fileChanges/${file.id}`);
-          }
-        }
-      }
-
-      const instrumentChanges = [];
-      const instrumentIdsToPurge = [];
-
-      const allInstrumentIds = new Set([
-        ...Object.keys(initialInstrumentQuantities || {}),
-        ...Object.keys(formState.instruments || {}),
-      ]);
-
-      for (const id of allInstrumentIds) {
-        const newQty = formState.instruments?.[id]?.quantity ?? 0;
-        const oldQty = initialInstrumentQuantities?.[id] ?? 0;
-        if (program && newQty === 0 && oldQty > 0) {
-          instrumentIdsToPurge.push(Number(id));
-          continue;
-        }
-        const instrumentDiff = newQty - oldQty;
-        if (instrumentDiff !== 0) {
-          instrumentChanges.push({
-            instrumentId: Number(id),
-            amountChanged: instrumentDiff,
-          });
-        }
-      }
-
-      if (program && instrumentIdsToPurge.length > 0) {
-        for (const instId of instrumentIdsToPurge) {
-          await deleteInstrumentChangesForProgramInstrument(
-            backend,
-            programId,
-            instId
-          );
-        }
-      }
-
-      const hasStudentChange = enrollmentDelta !== 0 || graduatedDelta !== 0;
-      const hasInstrumentChange = instrumentChanges.length > 0;
-      const hasMediaChange = pendingMedia.length > 0 || pendingFiles.length > 0;
-
-      const isNewProgram = !program;
-
-      if (
-        isNewProgram ||
-        hasStudentChange ||
-        hasInstrumentChange ||
-        hasMediaChange
-      ) {
-        const updateResponse = await backend.post(`/program-updates`, {
-          title: isNewProgram ? 'Program Created' : 'update program stats',
-          program_id: programId,
-          created_by: currentUser?.uid || currentUser?.id,
-          update_date: new Date().toISOString(),
-          note: isNewProgram ? 'Program Created' : 'Program update',
-          show_on_table: isNewProgram,
-        });
-
-        const updateId = updateResponse.data.id;
-        if (hasStudentChange) {
-          await backend.post(`/enrollmentChange`, {
-            update_id: updateId,
-            enrollment_change: enrollmentDelta,
-            graduated_change: graduatedDelta,
-            event_type: 'other',
-          });
-        }
-        if (hasInstrumentChange) {
-          for (const instrumentChange of instrumentChanges) {
-            await backend.post(`/instrument-changes`, {
-              instrumentId: instrumentChange.instrumentId,
-              updateId,
-              amountChanged: instrumentChange.amountChanged,
-              event_type: 'other',
-            });
-          }
-        }
-        if (hasMediaChange) {
-          for (const fileChange of pendingFiles) {
-            await backend.post(`/fileChanges`, {
-              update_id: updateId,
-              s3_key: fileChange.s3_key,
-              file_name: fileChange.file_name,
-              file_type: fileChange.file_type,
-            });
-          }
-          for (const mediaChange of pendingMedia) {
-            await backend.post(`/mediaChange`, {
-              update_id: updateId,
-              s3_key: mediaChange.s3_key,
-              file_name: mediaChange.file_name,
-              file_type: mediaChange.file_type,
-              is_thumbnail: false,
-            });
-          }
-        }
-      }
-
-      onSave?.();
-      onClose();
+      await saveProgramForm({
+        backend,
+        currentUser,
+        program,
+        formState,
+        initialProgramDirectorIds,
+        setInitialProgramDirectorIds,
+        initialInstrumentQuantities,
+        initialCurriculumLinks,
+        initialGraduated,
+        onSave,
+        onClose,
+      });
     } catch (err) {
       console.error('Error saving program:', err);
     }
@@ -794,180 +241,33 @@ export const ProgramForm = ({
             </HStack>
 
             {activeTab === 'overview' && (
-              <>
-                <Box>
-                  <Heading
-                    size="md"
-                    fontWeight="semibold"
-                    mb={3}
-                  >
-                    {t('programForm.generalInformation')}
-                  </Heading>
-                  <VStack
-                    align="stretch"
-                    spacing={4}
-                  >
-                    <FormControl isRequired>
-                      <FormLabel
-                        size="sm"
-                        fontWeight="normal"
-                        color="gray"
-                      >
-                        {t('programForm.programName')}
-                      </FormLabel>
-                      <Input
-                        placeholder={t('programForm.enterProgramTitle')}
-                        value={formState.programName || ''}
-                        onChange={(e) =>
-                          handleProgramNameChange(e.target.value)
-                        }
-                      />
-                    </FormControl>
-
-                    <PartnerOrganizationField
-                      label={t('programForm.partnerOrgName')}
-                      valueId={formState.partnerOrg}
-                      onChangeId={(id) =>
-                        setFormState((prev) => ({
-                          ...prev,
-                          partnerOrg: id,
-                        }))
-                      }
-                    />
-
-                    <FormControl>
-                      <Checkbox
-                        isChecked={Boolean(formState.showPartnerOrgOnMap)}
-                        onChange={(e) =>
-                          setFormState((prev) => ({
-                            ...prev,
-                            showPartnerOrgOnMap: e.target.checked,
-                          }))
-                        }
-                      >
-                        {t('programForm.showPartnerOnMap')}
-                      </Checkbox>
-                      {/* TODO: Implement persistence and map behavior for showPartnerOrgOnMap (API + map layer). */}
-                    </FormControl>
-
-                    <FormControl isRequired>
-                      <FormLabel
-                        size="sm"
-                        fontWeight="normal"
-                        color="gray"
-                      >
-                        {t('programForm.status')}
-                      </FormLabel>
-                      <Select
-                        value={formState.status ?? ''}
-                        onChange={(e) =>
-                          handleProgramStatusChange(e.target.value)
-                        }
-                        placeholder={t('programForm.selectStatus')}
-                      >
-                        <option value="Active">
-                          {t('programForm.launched')}
-                        </option>
-                        <option value="Inactive">
-                          {t('programForm.developing')}
-                        </option>
-                      </Select>
-                    </FormControl>
-
-                    <FormControl>
-                      <FormLabel
-                        size="sm"
-                        fontWeight="normal"
-                        color="gray"
-                      >
-                        {t('programForm.launchDate')}
-                      </FormLabel>
-                      <Input
-                        type="date"
-                        placeholder={t('programForm.datePlaceholder')}
-                        value={formState.launchDate || ''}
-                        onChange={(e) =>
-                          handleProgramLaunchDateChange(e.target.value)
-                        }
-                      />
-                    </FormControl>
-                  </VStack>
-                </Box>
-
-                <LocationLanguageSection
-                  formState={formState}
-                  setFormData={setFormState}
-                  languageOptions={languageOptions}
-                  onLanguagesChange={handleLanguageChange}
-                />
-
-                <StudentsInstrumentsSection
-                  formState={formState}
-                  setFormData={setFormState}
-                />
-
-                <AssignedDirectorsSection
-                  regionId={formState.regionId}
-                  formState={formState}
-                  setFormData={setFormState}
-                />
-
-                <ResourcesSection
-                  formState={formState}
-                  setFormData={setFormState}
-                  programId={program?.id}
-                  backend={backend}
-                  onOpenMediaModal={() => {
-                    mediaUploadTargetRef.current = 'files';
-                    mediaUploadModal.onOpen();
-                  }}
-                  onSeeAllMedia={() => setActiveTab('media')}
-                />
-              </>
+              <ProgramFormOverviewTab
+                formState={formState}
+                setFormState={setFormState}
+                languageOptions={languageOptions}
+                onProgramNameChange={handleProgramNameChange}
+                onProgramStatusChange={handleProgramStatusChange}
+                onProgramLaunchDateChange={handleProgramLaunchDateChange}
+                onLanguageChange={handleLanguageChange}
+                programId={program?.id}
+                backend={backend}
+                onOpenMediaModal={() => {
+                  mediaUploadTargetRef.current = 'files';
+                  mediaUploadModal.onOpen();
+                }}
+                onSeeAllMedia={() => setActiveTab('media')}
+              />
             )}
 
             {activeTab === 'media' && (
-              <VStack
-                align="stretch"
-                spacing={4}
-              >
-                <Heading
-                  size="md"
-                  fontWeight="semibold"
-                >
-                  {t('programForm.mediaHeading')}
-                </Heading>
-                <Button
-                  size="sm"
-                  variant="outline"
-                  onClick={() => {
-                    mediaUploadTargetRef.current = 'media';
-                    mediaUploadModal.onOpen();
-                  }}
-                >
-                  {t('common.add')}
-                </Button>
-                <HStack
-                  wrap="wrap"
-                  spacing={3}
-                >
-                  {(formState.media ?? [])
-                    .filter((item) => !isPdfByType(item))
-                    .map((item, index) => (
-                      <MediaPreviewTag
-                        key={item.id || item.s3_key || `media-${index}`}
-                        item={item}
-                        onRemove={() => {
-                          setFormState((prev) => ({
-                            ...prev,
-                            media: removeFormItemByIdOrKey(prev.media, item),
-                          }));
-                        }}
-                        isMedia={true}
-                      />
-                    ))}
-                </HStack>
-              </VStack>
+              <ProgramFormMediaTab
+                formState={formState}
+                setFormState={setFormState}
+                onOpenMediaUpload={() => {
+                  mediaUploadTargetRef.current = 'media';
+                  mediaUploadModal.onOpen();
+                }}
+              />
             )}
           </VStack>
         </DrawerBody>
