+++ client/src/components/media/Media.jsx
@@ -5,7 +5,6 @@ import {
   Box,
   Button,
   Center,
-  Container,
   Heading,
   HStack,
   Input,
